Implemented switch-page on OS/2, and changed the signal prototype to unsigned long instead of int.  Also dw_notebook_page_query should return unsigned long as well.
git-svn-id: https://svn.netlabs.org/repos/dwindows/trunk@348 81767d24-ef19-dc11-ae90-00e081727c95

+++ dw.h
@@ -875,7 +875,7 @@ void API dw_notebook_page_destroy(HWND handle, unsigned int pageid);
 void API dw_notebook_page_set_text(HWND handle, unsigned long pageid, char *text);
 void API dw_notebook_page_set_status_text(HWND handle, unsigned long pageid, char *text);
 void API dw_notebook_page_set(HWND handle, unsigned int pageid);
-unsigned int API dw_notebook_page_query(HWND handle);
+unsigned long API dw_notebook_page_query(HWND handle);
 void API dw_notebook_pack(HWND handle, unsigned long pageid, HWND page);
 HWND API dw_splitbar_new(int type, HWND topleft, HWND bottomright, unsigned long id);
 void API dw_splitbar_set(HWND handle, float percent);
